Updating docs, logging and fixing CLI.
The usage of some of the elpm functions has changed, so updating the
readme to account for that. Further, the last change to move the
straight.el portion of the package into an emacs subprocess created some
issues. First, the logging was lost, which isn't that big of a deal, but
when using the CLI it's comforting to see that stuff is being installed.
This change restores the logging using the a process filter. The second
issue introduced is that the subprocess is asynchronous, which impacts
the functioning of the CLI. Since the main emacs process has nothing to
do it exited immediately as a result not allowing the subprocess to
finish. To fix this we added a sleep loop.